pyproject.toml
[project]
name = "yatra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
]
